feat: update to use hypercerts api v2
