+++ app/components/create-listing-dialog/listing-progress.tsx
@@ -1,8 +1,8 @@
 "use client";
-import type { FullHypercert } from "@/app/graphql-queries/hypercerts";
 import { Button } from "@/components/ui/button";
 import { DIVVI_DATA_SUFFIX } from "@/config/divvi";
 import { GAINFOREST_TIP_ADDRESS, GAINFOREST_TIP_AMOUNT } from "@/config/tip";
+import type { FullHypercert } from "@/graphql/hypercerts/queries/hypercerts";
 import { useHypercertClient } from "@/hooks/use-hypercerts-client";
 import { cn } from "@/lib/utils";
 import { submitReferral } from "@divvi/referral-sdk";
