+++ app/components/unlist-dialog.tsx
@@ -16,7 +16,7 @@ import { CircleAlert, CircleCheck, Loader2 } from "lucide-react";
 import type React from "react";
 import { useCallback, useMemo, useState } from "react";
 import { useAccount } from "wagmi";
-import { fetchHypercertById } from "../graphql-queries/hypercerts";
+import { fetchHypercertById } from "../../graphql/hypercerts/queries/hypercerts";
 import { catchError } from "../utils";
 
 // Status type definition
