+++ app/contexts/full-hypercert.tsx
@@ -1,5 +1,5 @@
 "use client";
-import type { FullHypercert } from "@/app/graphql-queries/hypercerts";
+import type { FullHypercert } from "@/graphql/hypercerts/queries/hypercerts";
 import { createContext, useContext } from "react";
 
 const FullHypercertContext = createContext<FullHypercert | null>(null);
