+++ app/hypercert/[hypercertId]/attestations/page.tsx
@@ -1,8 +1,8 @@
 import Progress from "@/app/components/shared/progress";
 import { FullHypercertProvider } from "@/app/contexts/full-hypercert";
-import { fetchFullHypercertById } from "@/app/graphql-queries/hypercerts";
 import { Button } from "@/components/ui/button";
 import { MotionWrapper } from "@/components/ui/motion-wrapper";
+import { fetchFullHypercertById } from "@/graphql/hypercerts/queries/hypercerts";
 import { ChevronLeft } from "lucide-react";
 import Link from "next/link";
 import AttestationsList from "./components/AttestationsList";
