+++ app/hypercert/[hypercertId]/components/FundingView.tsx
@@ -4,11 +4,11 @@ import usePriceFeed from "@/app/PriceFeedProvider";
 import CreateListingDialog from "@/app/components/create-listing-dialog";
 import Progress from "@/app/components/shared/progress";
 import useFullHypercert from "@/app/contexts/full-hypercert";
-import type { FullHypercert } from "@/app/graphql-queries/hypercerts";
 import { Button } from "@/components/ui/button";
 import { useModal } from "@/components/ui/modal/context";
 import QuickTooltip from "@/components/ui/quicktooltip";
 import { RAW_TOKENS_CONFIG } from "@/config/raw-tokens";
+import type { FullHypercert } from "@/graphql/hypercerts/queries/hypercerts";
 import { calculateBigIntPercentage } from "@/lib/calculateBigIntPercentage";
 import { convertCurrencyPriceToUSD, formatDecimals } from "@/lib/utils";
 import { cn } from "@/lib/utils";
