+++ graphql/hypercerts/env.d.ts
@@ -1,0 +1,128 @@
+/* eslint-disable */
+/* prettier-ignore */
+
+export type introspection_types = {
+    'AllowlistRecord': { kind: 'OBJECT'; name: 'AllowlistRecord'; fields: { 'claimed': { name: 'claimed'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'entry': { name: 'entry'; type: { kind: 'SCALAR'; name: 'Float'; ofType: null; } }; 'hypercert': { name: 'hypercert'; type: { kind: 'OBJECT'; name: 'Hypercert'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'leaf': { name: 'leaf'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'proof': { name: 'proof'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; } }; 'root': { name: 'root'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'token_id': { name: 'token_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'total_units': { name: 'total_units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'units': { name: 'units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'user_address': { name: 'user_address'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; }; };
+    'AllowlistRecordHypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'AllowlistRecordHypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'AllowlistRecordSortOptions': { kind: 'INPUT_OBJECT'; name: 'AllowlistRecordSortOptions'; isOneOf: false; inputFields: [{ name: 'hypercert_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'token_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'leaf'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'entry'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'user_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'claimed'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'proof'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'units'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'total_units'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'root'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'AllowlistRecordWhereInput': { kind: 'INPUT_OBJECT'; name: 'AllowlistRecordWhereInput'; isOneOf: false; inputFields: [{ name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'leaf'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'entry'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'user_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'claimed'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'proof'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'total_units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'root'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert'; type: { kind: 'INPUT_OBJECT'; name: 'AllowlistRecordHypercertWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'Attestation': { kind: 'OBJECT'; name: 'Attestation'; fields: { 'attester': { name: 'attester'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'creation_block_number': { name: 'creation_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creation_block_timestamp': { name: 'creation_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'SCALAR'; name: 'JSON'; ofType: null; } }; 'eas_schema': { name: 'eas_schema'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'AttestationSchemaBaseType'; ofType: null; }; } }; 'hypercert': { name: 'hypercert'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'HypercertBaseType'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'last_update_block_number': { name: 'last_update_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'last_update_block_timestamp': { name: 'last_update_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'metadata': { name: 'metadata'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Metadata'; ofType: null; }; } }; 'recipient': { name: 'recipient'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'schema_uid': { name: 'schema_uid'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'uid': { name: 'uid'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; }; };
+    'AttestationAttestationSchemaWhereInput': { kind: 'INPUT_OBJECT'; name: 'AttestationAttestationSchemaWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uid'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'resolver'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'revocable'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'AttestationHypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'AttestationHypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'AttestationSchema': { kind: 'OBJECT'; name: 'AttestationSchema'; fields: { 'attestations': { name: 'attestations'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetAttestationsResponse'; ofType: null; }; } }; 'chain_id': { name: 'chain_id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'resolver': { name: 'resolver'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'revocable': { name: 'revocable'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'schema': { name: 'schema'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'uid': { name: 'uid'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ID'; ofType: null; }; } }; }; };
+    'AttestationSchemaAttestationWhereInput': { kind: 'INPUT_OBJECT'; name: 'AttestationSchemaAttestationWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uid'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attester'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'recipient'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'resolver'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'supported_schemas_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contract_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'AttestationSchemaBaseType': { kind: 'OBJECT'; name: 'AttestationSchemaBaseType'; fields: { 'chain_id': { name: 'chain_id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'resolver': { name: 'resolver'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'revocable': { name: 'revocable'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'schema': { name: 'schema'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'uid': { name: 'uid'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ID'; ofType: null; }; } }; }; };
+    'AttestationSchemaSortOptions': { kind: 'INPUT_OBJECT'; name: 'AttestationSchemaSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'chain_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'uid'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'resolver'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'revocable'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'AttestationSchemaWhereInput': { kind: 'INPUT_OBJECT'; name: 'AttestationSchemaWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uid'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'resolver'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'revocable'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations'; type: { kind: 'INPUT_OBJECT'; name: 'AttestationSchemaAttestationWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'AttestationSortOptions': { kind: 'INPUT_OBJECT'; name: 'AttestationSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'uid'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'last_update_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'last_update_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'attester'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'recipient'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'resolver'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'supported_schemas_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'contract_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'AttestationWhereInput': { kind: 'INPUT_OBJECT'; name: 'AttestationWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uid'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attester'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'recipient'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'resolver'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'supported_schemas_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contract_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert'; type: { kind: 'INPUT_OBJECT'; name: 'AttestationHypercertWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'eas_schema'; type: { kind: 'INPUT_OBJECT'; name: 'AttestationAttestationSchemaWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'BigInt': unknown;
+    'BigIntSearchOptions': { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; isOneOf: false; inputFields: [{ name: 'eq'; type: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; defaultValue: null }, { name: 'gt'; type: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; defaultValue: null }, { name: 'gte'; type: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; defaultValue: null }, { name: 'lt'; type: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; defaultValue: null }, { name: 'lte'; type: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; defaultValue: null }]; };
+    'Blueprint': { kind: 'OBJECT'; name: 'Blueprint'; fields: { 'admins': { name: 'admins'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'User'; ofType: null; }; }; }; } }; 'created_at': { name: 'created_at'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'form_values': { name: 'form_values'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'JSON'; ofType: null; }; } }; 'hypercerts': { name: 'hypercerts'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'HypercertsResponse'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'minted': { name: 'minted'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'minter_address': { name: 'minter_address'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; }; };
+    'BlueprintSortOptions': { kind: 'INPUT_OBJECT'; name: 'BlueprintSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'created_at'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'minter_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'minted'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'admin_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'BlueprintUserWhereInput': { kind: 'INPUT_OBJECT'; name: 'BlueprintUserWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'display_name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'BlueprintWhereInput': { kind: 'INPUT_OBJECT'; name: 'BlueprintWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'created_at'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'minter_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'minted'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'admin_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'admins'; type: { kind: 'INPUT_OBJECT'; name: 'BlueprintUserWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'Boolean': unknown;
+    'BooleanSearchOptions': { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; isOneOf: false; inputFields: [{ name: 'eq'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; defaultValue: null }]; };
+    'Collection': { kind: 'OBJECT'; name: 'Collection'; fields: { 'admins': { name: 'admins'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'User'; ofType: null; }; }; }; } }; 'blueprints': { name: 'blueprints'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Blueprint'; ofType: null; }; }; } }; 'chain_ids': { name: 'chain_ids'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; }; } }; 'created_at': { name: 'created_at'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'description': { name: 'description'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'hypercerts': { name: 'hypercerts'; type: { kind: 'OBJECT'; name: 'HypercertsResponse'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'name': { name: 'name'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; }; };
+    'CollectionBlueprintWhereInput': { kind: 'INPUT_OBJECT'; name: 'CollectionBlueprintWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'created_at'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'minter_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'minted'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'admin_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'CollectionHypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'CollectionHypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'CollectionSortOptions': { kind: 'INPUT_OBJECT'; name: 'CollectionSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'name'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'description'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'created_at'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'CollectionUserWhereInput': { kind: 'INPUT_OBJECT'; name: 'CollectionUserWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'display_name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'CollectionWhereInput': { kind: 'INPUT_OBJECT'; name: 'CollectionWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'description'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'created_at'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'admins'; type: { kind: 'INPUT_OBJECT'; name: 'CollectionUserWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'hypercerts'; type: { kind: 'INPUT_OBJECT'; name: 'CollectionHypercertWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'blueprints'; type: { kind: 'INPUT_OBJECT'; name: 'CollectionBlueprintWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'Contract': { kind: 'OBJECT'; name: 'Contract'; fields: { 'chain_id': { name: 'chain_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'contract_address': { name: 'contract_address'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'start_block': { name: 'start_block'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; }; };
+    'ContractSortOptions': { kind: 'INPUT_OBJECT'; name: 'ContractSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'contract_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'chain_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'ContractWhereInput': { kind: 'INPUT_OBJECT'; name: 'ContractWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contract_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'EthBigInt': unknown;
+    'Float': unknown;
+    'Fraction': { kind: 'OBJECT'; name: 'Fraction'; fields: { 'burned': { name: 'burned'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'claims_id': { name: 'claims_id'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'creation_block_number': { name: 'creation_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creation_block_timestamp': { name: 'creation_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'fraction_id': { name: 'fraction_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'last_update_block_number': { name: 'last_update_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'last_update_block_timestamp': { name: 'last_update_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'metadata': { name: 'metadata'; type: { kind: 'OBJECT'; name: 'Metadata'; ofType: null; } }; 'orders': { name: 'orders'; type: { kind: 'OBJECT'; name: 'GetOrdersResponse'; ofType: null; } }; 'owner_address': { name: 'owner_address'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'sales': { name: 'sales'; type: { kind: 'OBJECT'; name: 'GetSalesResponse'; ofType: null; } }; 'token_id': { name: 'token_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'units': { name: 'units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; }; };
+    'FractionMetadataWhereInput': { kind: 'INPUT_OBJECT'; name: 'FractionMetadataWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'description'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'allow_list_uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contributors'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'external_url'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_scope'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'rights'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_scope'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_timeframe_from'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_timeframe_to'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_timeframe_from'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_timeframe_to'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'FractionSortOptions': { kind: 'INPUT_OBJECT'; name: 'FractionSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'last_update_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'last_update_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'owner_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'units'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'hypercert_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'fraction_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'token_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'burned'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'FractionWhereInput': { kind: 'INPUT_OBJECT'; name: 'FractionWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'owner_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'fraction_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'metadata'; type: { kind: 'INPUT_OBJECT'; name: 'FractionMetadataWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'GetAllowlistRecordResponse': { kind: 'OBJECT'; name: 'GetAllowlistRecordResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'AllowlistRecord'; ofType: null; }; }; } }; }; };
+    'GetAttestationsResponse': { kind: 'OBJECT'; name: 'GetAttestationsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Attestation'; ofType: null; }; }; } }; }; };
+    'GetAttestationsSchemaResponse': { kind: 'OBJECT'; name: 'GetAttestationsSchemaResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'AttestationSchema'; ofType: null; }; }; } }; }; };
+    'GetBlueprintsResponse': { kind: 'OBJECT'; name: 'GetBlueprintsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Blueprint'; ofType: null; }; }; } }; }; };
+    'GetCollectionsResponse': { kind: 'OBJECT'; name: 'GetCollectionsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Collection'; ofType: null; }; }; } }; }; };
+    'GetContractsResponse': { kind: 'OBJECT'; name: 'GetContractsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Contract'; ofType: null; }; }; } }; }; };
+    'GetFractionsResponse': { kind: 'OBJECT'; name: 'GetFractionsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Fraction'; ofType: null; }; }; } }; }; };
+    'GetHyperboardOwnersResponse': { kind: 'OBJECT'; name: 'GetHyperboardOwnersResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'HyperboardOwner'; ofType: null; }; }; } }; }; };
+    'GetHyperboardsResponse': { kind: 'OBJECT'; name: 'GetHyperboardsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Hyperboard'; ofType: null; }; }; } }; }; };
+    'GetHypercertsResponse': { kind: 'OBJECT'; name: 'GetHypercertsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Hypercert'; ofType: null; }; }; } }; }; };
+    'GetMetadataResponse': { kind: 'OBJECT'; name: 'GetMetadataResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Metadata'; ofType: null; }; }; } }; }; };
+    'GetOrdersForHypercertResponse': { kind: 'OBJECT'; name: 'GetOrdersForHypercertResponse'; fields: { 'cheapestOrder': { name: 'cheapestOrder'; type: { kind: 'OBJECT'; name: 'Order'; ofType: null; } }; 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Order'; ofType: null; }; }; } }; 'totalUnitsForSale': { name: 'totalUnitsForSale'; type: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; } }; }; };
+    'GetOrdersResponse': { kind: 'OBJECT'; name: 'GetOrdersResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Order'; ofType: null; }; }; } }; }; };
+    'GetSalesResponse': { kind: 'OBJECT'; name: 'GetSalesResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Sale'; ofType: null; }; }; } }; }; };
+    'GetSectionEntryOwnersResponse': { kind: 'OBJECT'; name: 'GetSectionEntryOwnersResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SectionEntryOwner'; ofType: null; }; }; } }; }; };
+    'GetSectionsResponse': { kind: 'OBJECT'; name: 'GetSectionsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Section'; ofType: null; }; }; } }; }; };
+    'GetSignatureRequestResponse': { kind: 'OBJECT'; name: 'GetSignatureRequestResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SignatureRequest'; ofType: null; }; }; } }; }; };
+    'GetUsersResponse': { kind: 'OBJECT'; name: 'GetUsersResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'User'; ofType: null; }; }; } }; }; };
+    'Hyperboard': { kind: 'OBJECT'; name: 'Hyperboard'; fields: { 'admins': { name: 'admins'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetUsersResponse'; ofType: null; }; } }; 'background_image': { name: 'background_image'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'chain_ids': { name: 'chain_ids'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; }; } }; 'grayscale_images': { name: 'grayscale_images'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'name': { name: 'name'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'owners': { name: 'owners'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetHyperboardOwnersResponse'; ofType: null; }; } }; 'sections': { name: 'sections'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetSectionsResponse'; ofType: null; }; } }; 'tile_border_color': { name: 'tile_border_color'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; }; };
+    'HyperboardCollectionWhereInput': { kind: 'INPUT_OBJECT'; name: 'HyperboardCollectionWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'description'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'created_at'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'HyperboardOwner': { kind: 'OBJECT'; name: 'HyperboardOwner'; fields: { 'address': { name: 'address'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'avatar': { name: 'avatar'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'chain_id': { name: 'chain_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'display_name': { name: 'display_name'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'percentage_owned': { name: 'percentage_owned'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'signature_requests': { name: 'signature_requests'; type: { kind: 'OBJECT'; name: 'GetSignatureRequestResponse'; ofType: null; } }; }; };
+    'HyperboardSortOptions': { kind: 'INPUT_OBJECT'; name: 'HyperboardSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'chain_ids'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'admin_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'HyperboardUserWhereInput': { kind: 'INPUT_OBJECT'; name: 'HyperboardUserWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'display_name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'HyperboardWhereInput': { kind: 'INPUT_OBJECT'; name: 'HyperboardWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_ids'; type: { kind: 'INPUT_OBJECT'; name: 'NumberArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'admin_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'collections'; type: { kind: 'INPUT_OBJECT'; name: 'HyperboardCollectionWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'admins'; type: { kind: 'INPUT_OBJECT'; name: 'HyperboardUserWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'Hypercert': { kind: 'OBJECT'; name: 'Hypercert'; fields: { 'attestations': { name: 'attestations'; type: { kind: 'OBJECT'; name: 'GetAttestationsResponse'; ofType: null; } }; 'attestations_count': { name: 'attestations_count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'burned': { name: 'burned'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'contract': { name: 'contract'; type: { kind: 'OBJECT'; name: 'Contract'; ofType: null; } }; 'contracts_id': { name: 'contracts_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'creation_block_number': { name: 'creation_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creation_block_timestamp': { name: 'creation_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creator_address': { name: 'creator_address'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'fractions': { name: 'fractions'; type: { kind: 'OBJECT'; name: 'GetFractionsResponse'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'last_update_block_number': { name: 'last_update_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'last_update_block_timestamp': { name: 'last_update_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'metadata': { name: 'metadata'; type: { kind: 'OBJECT'; name: 'Metadata'; ofType: null; } }; 'orders': { name: 'orders'; type: { kind: 'OBJECT'; name: 'GetOrdersForHypercertResponse'; ofType: null; } }; 'sales': { name: 'sales'; type: { kind: 'OBJECT'; name: 'GetSalesResponse'; ofType: null; } }; 'sales_count': { name: 'sales_count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'token_id': { name: 'token_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'units': { name: 'units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'uri': { name: 'uri'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; }; };
+    'HypercertAttestationWhereInput': { kind: 'INPUT_OBJECT'; name: 'HypercertAttestationWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uid'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attester'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'recipient'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'resolver'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'supported_schemas_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contract_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'HypercertBaseType': { kind: 'OBJECT'; name: 'HypercertBaseType'; fields: { 'attestations_count': { name: 'attestations_count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'burned': { name: 'burned'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'contracts_id': { name: 'contracts_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'creation_block_number': { name: 'creation_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creation_block_timestamp': { name: 'creation_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creator_address': { name: 'creator_address'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'last_update_block_number': { name: 'last_update_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'last_update_block_timestamp': { name: 'last_update_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'sales_count': { name: 'sales_count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'token_id': { name: 'token_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'units': { name: 'units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'uri': { name: 'uri'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; }; };
+    'HypercertContractWhereInput': { kind: 'INPUT_OBJECT'; name: 'HypercertContractWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contract_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'HypercertFractionWhereInput': { kind: 'INPUT_OBJECT'; name: 'HypercertFractionWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'owner_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'fraction_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'HypercertMetadataWhereInput': { kind: 'INPUT_OBJECT'; name: 'HypercertMetadataWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'description'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'allow_list_uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contributors'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'external_url'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_scope'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'rights'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_scope'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_timeframe_from'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_timeframe_to'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_timeframe_from'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_timeframe_to'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'HypercertSortOptions': { kind: 'INPUT_OBJECT'; name: 'HypercertSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'hypercert_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creator_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'token_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'units'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'last_update_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'last_update_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'sales_count'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'attestations_count'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'uri'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'burned'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'HypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'HypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contract'; type: { kind: 'INPUT_OBJECT'; name: 'HypercertContractWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'metadata'; type: { kind: 'INPUT_OBJECT'; name: 'HypercertMetadataWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'attestations'; type: { kind: 'INPUT_OBJECT'; name: 'HypercertAttestationWhereInput'; ofType: null; }; defaultValue: "{}" }, { name: 'fractions'; type: { kind: 'INPUT_OBJECT'; name: 'HypercertFractionWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'HypercertWithMetadata': { kind: 'OBJECT'; name: 'HypercertWithMetadata'; fields: { 'attestations_count': { name: 'attestations_count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'burned': { name: 'burned'; type: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; } }; 'contracts_id': { name: 'contracts_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'creation_block_number': { name: 'creation_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creation_block_timestamp': { name: 'creation_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creator_address': { name: 'creator_address'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'last_update_block_number': { name: 'last_update_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'last_update_block_timestamp': { name: 'last_update_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'metadata': { name: 'metadata'; type: { kind: 'OBJECT'; name: 'Metadata'; ofType: null; } }; 'sales_count': { name: 'sales_count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'token_id': { name: 'token_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'units': { name: 'units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'uri': { name: 'uri'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; }; };
+    'HypercertsResponse': { kind: 'OBJECT'; name: 'HypercertsResponse'; fields: { 'count': { name: 'count'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; } }; 'data': { name: 'data'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'HypercertBaseType'; ofType: null; }; }; } }; }; };
+    'ID': unknown;
+    'Int': unknown;
+    'JSON': unknown;
+    'Metadata': { kind: 'OBJECT'; name: 'Metadata'; fields: { 'allow_list_uri': { name: 'allow_list_uri'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'contributors': { name: 'contributors'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; } }; 'description': { name: 'description'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'external_url': { name: 'external_url'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'image': { name: 'image'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'impact_scope': { name: 'impact_scope'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; } }; 'impact_timeframe_from': { name: 'impact_timeframe_from'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'impact_timeframe_to': { name: 'impact_timeframe_to'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'name': { name: 'name'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'properties': { name: 'properties'; type: { kind: 'SCALAR'; name: 'JSON'; ofType: null; } }; 'rights': { name: 'rights'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; } }; 'uri': { name: 'uri'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'work_scope': { name: 'work_scope'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; } }; 'work_timeframe_from': { name: 'work_timeframe_from'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'work_timeframe_to': { name: 'work_timeframe_to'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; }; };
+    'MetadataHypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'MetadataHypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'MetadataSortOptions': { kind: 'INPUT_OBJECT'; name: 'MetadataSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'name'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'description'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'uri'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'allow_list_uri'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'contributors'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'external_url'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'impact_scope'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'rights'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'work_scope'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'work_timeframe_from'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'work_timeframe_to'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'impact_timeframe_from'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'impact_timeframe_to'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'MetadataWhereInput': { kind: 'INPUT_OBJECT'; name: 'MetadataWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'description'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'allow_list_uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'contributors'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'external_url'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_scope'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'rights'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_scope'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_timeframe_from'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'work_timeframe_to'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_timeframe_from'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'impact_timeframe_to'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert'; type: { kind: 'INPUT_OBJECT'; name: 'MetadataHypercertWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'NumberArraySearchOptions': { kind: 'INPUT_OBJECT'; name: 'NumberArraySearchOptions'; isOneOf: false; inputFields: [{ name: 'arrayContains'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; }; }; defaultValue: null }, { name: 'arrayOverlaps'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BigInt'; ofType: null; }; }; }; defaultValue: null }]; };
+    'NumberSearchOptions': { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; isOneOf: false; inputFields: [{ name: 'eq'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; defaultValue: null }, { name: 'in'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; }; }; defaultValue: null }, { name: 'gt'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; defaultValue: null }, { name: 'gte'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; defaultValue: null }, { name: 'lt'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; defaultValue: null }, { name: 'lte'; type: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; defaultValue: null }]; };
+    'Order': { kind: 'OBJECT'; name: 'Order'; fields: { 'additionalParameters': { name: 'additionalParameters'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'amounts': { name: 'amounts'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; }; }; } }; 'chainId': { name: 'chainId'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; } }; 'collection': { name: 'collection'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'collectionType': { name: 'collectionType'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'currency': { name: 'currency'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'endTime': { name: 'endTime'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'globalNonce': { name: 'globalNonce'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'hypercert': { name: 'hypercert'; type: { kind: 'OBJECT'; name: 'HypercertWithMetadata'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'invalidated': { name: 'invalidated'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'itemIds': { name: 'itemIds'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; }; } }; 'orderNonce': { name: 'orderNonce'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'price': { name: 'price'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'pricePerPercentInToken': { name: 'pricePerPercentInToken'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'pricePerPercentInUSD': { name: 'pricePerPercentInUSD'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'quoteType': { name: 'quoteType'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'signature': { name: 'signature'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'signer': { name: 'signer'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'startTime': { name: 'startTime'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'strategyId': { name: 'strategyId'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'subsetNonce': { name: 'subsetNonce'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'validator_codes': { name: 'validator_codes'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Int'; ofType: null; }; }; } }; }; };
+    'OrderHypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'OrderHypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'OrderSortOptions': { kind: 'INPUT_OBJECT'; name: 'OrderSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'hypercert_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'createdAt'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'quoteType'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'globalNonce'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'orderNonce'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'strategyId'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'collectionType'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'collection'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'currency'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'signer'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'startTime'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'endTime'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'price'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'chainId'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'subsetNonce'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'itemIds'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'amounts'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'invalidated'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'OrderWhereInput': { kind: 'INPUT_OBJECT'; name: 'OrderWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'createdAt'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'quoteType'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'globalNonce'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'orderNonce'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'strategyId'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'collectionType'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'collection'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'currency'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'signer'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'startTime'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'endTime'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'price'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chainId'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'subsetNonce'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'itemIds'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'amounts'; type: { kind: 'INPUT_OBJECT'; name: 'NumberArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'invalidated'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert'; type: { kind: 'INPUT_OBJECT'; name: 'OrderHypercertWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'Query': { kind: 'OBJECT'; name: 'Query'; fields: { 'allowlistRecords': { name: 'allowlistRecords'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetAllowlistRecordResponse'; ofType: null; }; } }; 'attestationSchemas': { name: 'attestationSchemas'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetAttestationsSchemaResponse'; ofType: null; }; } }; 'attestations': { name: 'attestations'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetAttestationsResponse'; ofType: null; }; } }; 'blueprints': { name: 'blueprints'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetBlueprintsResponse'; ofType: null; }; } }; 'collections': { name: 'collections'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetCollectionsResponse'; ofType: null; }; } }; 'contracts': { name: 'contracts'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetContractsResponse'; ofType: null; }; } }; 'fractions': { name: 'fractions'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetFractionsResponse'; ofType: null; }; } }; 'hyperboards': { name: 'hyperboards'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetHyperboardsResponse'; ofType: null; }; } }; 'hypercerts': { name: 'hypercerts'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetHypercertsResponse'; ofType: null; }; } }; 'metadata': { name: 'metadata'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetMetadataResponse'; ofType: null; }; } }; 'orders': { name: 'orders'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetOrdersResponse'; ofType: null; }; } }; 'sales': { name: 'sales'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetSalesResponse'; ofType: null; }; } }; 'signatureRequests': { name: 'signatureRequests'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetSignatureRequestResponse'; ofType: null; }; } }; 'users': { name: 'users'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetUsersResponse'; ofType: null; }; } }; }; };
+    'Sale': { kind: 'OBJECT'; name: 'Sale'; fields: { 'amounts': { name: 'amounts'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; }; } }; 'buyer': { name: 'buyer'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'collection': { name: 'collection'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'creation_block_number': { name: 'creation_block_number'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'creation_block_timestamp': { name: 'creation_block_timestamp'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'currency': { name: 'currency'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'currency_amount': { name: 'currency_amount'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; } }; 'hypercert': { name: 'hypercert'; type: { kind: 'OBJECT'; name: 'HypercertWithMetadata'; ofType: null; } }; 'hypercert_id': { name: 'hypercert_id'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'item_ids': { name: 'item_ids'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; }; } }; 'seller': { name: 'seller'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'strategy_id': { name: 'strategy_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'transaction_hash': { name: 'transaction_hash'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; }; };
+    'SaleHypercertWhereInput': { kind: 'INPUT_OBJECT'; name: 'SaleHypercertWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creator_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'token_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'units'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'last_update_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'sales_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'attestations_count'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'uri'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'burned'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'SaleSortOptions': { kind: 'INPUT_OBJECT'; name: 'SaleSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'buyer'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'seller'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'strategy_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'currency'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'collection'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'item_ids'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'hypercert_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'amounts'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'transaction_hash'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_number'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'creation_block_timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'SaleWhereInput': { kind: 'INPUT_OBJECT'; name: 'SaleWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'buyer'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'seller'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'strategy_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'currency'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'collection'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'item_ids'; type: { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert_id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'amounts'; type: { kind: 'INPUT_OBJECT'; name: 'NumberArraySearchOptions'; ofType: null; }; defaultValue: null }, { name: 'transaction_hash'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_number'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'creation_block_timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'hypercert'; type: { kind: 'INPUT_OBJECT'; name: 'SaleHypercertWhereInput'; ofType: null; }; defaultValue: "{}" }]; };
+    'Section': { kind: 'OBJECT'; name: 'Section'; fields: { 'collections': { name: 'collections'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Collection'; ofType: null; }; }; }; } }; 'entries': { name: 'entries'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'SectionEntry'; ofType: null; }; }; }; } }; 'label': { name: 'label'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'owners': { name: 'owners'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetHyperboardOwnersResponse'; ofType: null; }; } }; }; };
+    'SectionEntry': { kind: 'OBJECT'; name: 'SectionEntry'; fields: { 'display_size': { name: 'display_size'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'is_blueprint': { name: 'is_blueprint'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Boolean'; ofType: null; }; } }; 'name': { name: 'name'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'owners': { name: 'owners'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'GetSectionEntryOwnersResponse'; ofType: null; }; } }; 'percentage_of_section': { name: 'percentage_of_section'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'total_units': { name: 'total_units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; }; };
+    'SectionEntryOwner': { kind: 'OBJECT'; name: 'SectionEntryOwner'; fields: { 'address': { name: 'address'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'avatar': { name: 'avatar'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'chain_id': { name: 'chain_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'display_name': { name: 'display_name'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'percentage': { name: 'percentage'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'Float'; ofType: null; }; } }; 'signature_requests': { name: 'signature_requests'; type: { kind: 'OBJECT'; name: 'GetSignatureRequestResponse'; ofType: null; } }; 'units': { name: 'units'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; }; };
+    'SignatureRequest': { kind: 'OBJECT'; name: 'SignatureRequest'; fields: { 'chain_id': { name: 'chain_id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; } }; 'message': { name: 'message'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'message_hash': { name: 'message_hash'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'purpose': { name: 'purpose'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'SignatureRequestPurpose'; ofType: null; }; } }; 'safe_address': { name: 'safe_address'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'status': { name: 'status'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'SignatureRequestStatus'; ofType: null; }; } }; 'timestamp': { name: 'timestamp'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; }; } }; }; };
+    'SignatureRequestPurpose': { name: 'SignatureRequestPurpose'; enumValues: 'UPDATE_USER_DATA'; };
+    'SignatureRequestSortOptions': { kind: 'INPUT_OBJECT'; name: 'SignatureRequestSortOptions'; isOneOf: false; inputFields: [{ name: 'safe_address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'message_hash'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'timestamp'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'chain_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'status'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'SignatureRequestStatus': { name: 'SignatureRequestStatus'; enumValues: 'PENDING' | 'EXECUTED' | 'CANCELED'; };
+    'SignatureRequestStatusSearchOptions': { kind: 'INPUT_OBJECT'; name: 'SignatureRequestStatusSearchOptions'; isOneOf: false; inputFields: [{ name: 'eq'; type: { kind: 'ENUM'; name: 'SignatureRequestStatus'; ofType: null; }; defaultValue: null }]; };
+    'SignatureRequestWhereInput': { kind: 'INPUT_OBJECT'; name: 'SignatureRequestWhereInput'; isOneOf: false; inputFields: [{ name: 'safe_address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'message_hash'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'timestamp'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'BigIntSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'status'; type: { kind: 'INPUT_OBJECT'; name: 'SignatureRequestStatusSearchOptions'; ofType: null; }; defaultValue: null }]; };
+    'SortOrder': { name: 'SortOrder'; enumValues: 'ascending' | 'descending'; };
+    'String': unknown;
+    'StringArraySearchOptions': { kind: 'INPUT_OBJECT'; name: 'StringArraySearchOptions'; isOneOf: false; inputFields: [{ name: 'arrayContains'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; }; defaultValue: null }, { name: 'arrayOverlaps'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; }; defaultValue: null }]; };
+    'StringSearchOptions': { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; isOneOf: false; inputFields: [{ name: 'eq'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; }; defaultValue: null }, { name: 'in'; type: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; }; defaultValue: null }, { name: 'contains'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; }; defaultValue: null }, { name: 'startsWith'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; }; defaultValue: null }, { name: 'endsWith'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; }; defaultValue: null }]; };
+    'User': { kind: 'OBJECT'; name: 'User'; fields: { 'address': { name: 'address'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'avatar': { name: 'avatar'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'chain_id': { name: 'chain_id'; type: { kind: 'SCALAR'; name: 'EthBigInt'; ofType: null; } }; 'display_name': { name: 'display_name'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'id': { name: 'id'; type: { kind: 'SCALAR'; name: 'ID'; ofType: null; } }; 'signature_requests': { name: 'signature_requests'; type: { kind: 'OBJECT'; name: 'GetSignatureRequestResponse'; ofType: null; } }; }; };
+    'UserSortOptions': { kind: 'INPUT_OBJECT'; name: 'UserSortOptions'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'address'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'display_name'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }, { name: 'chain_id'; type: { kind: 'ENUM'; name: 'SortOrder'; ofType: null; }; defaultValue: "null" }]; };
+    'UserWhereInput': { kind: 'INPUT_OBJECT'; name: 'UserWhereInput'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'address'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'display_name'; type: { kind: 'INPUT_OBJECT'; name: 'StringSearchOptions'; ofType: null; }; defaultValue: null }, { name: 'chain_id'; type: { kind: 'INPUT_OBJECT'; name: 'NumberSearchOptions'; ofType: null; }; defaultValue: null }]; };
+};
+
+/** An IntrospectionQuery representation of your schema.
+ *
+ * @remarks
+ * This is an introspection of your schema saved as a file by GraphQLSP.
+ * It will automatically be used by `gql.tada` to infer the types of your GraphQL documents.
+ * If you need to reuse this data or update your `scalars`, update `tadaOutputLocation` to
+ * instead save to a .ts instead of a .d.ts file.
+ */
+export type introspection = {
+  name: 'hypercerts';
+  query: 'Query';
+  mutation: never;
+  subscription: never;
+  types: introspection_types;
+};
+
+import * as gqlTada from 'gql.tada';
