+++ graphql/hypercerts/queries/templates.ts
