+++ tsconfig.json
@@ -19,9 +19,14 @@
         "name": "next"
       },
       {
-        "name": "@0no-co/graphqlsp",
-        "schema": "https://staging-api.hypercerts.org/v1/graphql",
-        "tadaOutputLocation": "./graphql-hypercerts-env.d.ts"
+        "name": "gql.tada/ts-plugin",
+        "schemas": [
+          {
+            "name": "hypercerts",
+            "schema": "https://api.hypercerts.org/v2/graphql",
+            "tadaOutputLocation": "./graphql/hypercerts/env.d.ts"
+          }
+        ]
       }
     ],
     "paths": {
